Add Bellman Ford Algorithm in C++
Given a weighted, directed, and connected graph of V vertices and E edges, Find the shortest distance of all the vertices from the source vertex S. If vertices can't be reached from the S then mark the distance as 10^8. Note: If the Graph contains a negative cycle then return an array consisting of only -1.